+++ src/main/java/com/lckp/jproxy/config/FilterConfig.java
@@ -9,14 +9,10 @@ import com.lckp.jproxy.filter.RadarrJackettFilter;
 import com.lckp.jproxy.filter.RadarrProwlarrFilter;
 import com.lckp.jproxy.filter.SonarrJackettFilter;
 import com.lckp.jproxy.filter.SonarrProwlarrFilter;
-import com.lckp.jproxy.filter.SonarrQbittorrentFilter;
-import com.lckp.jproxy.filter.SonarrTransmissionFilter;
 import com.lckp.jproxy.service.IRadarrJackettService;
 import com.lckp.jproxy.service.IRadarrProwlarrService;
 import com.lckp.jproxy.service.ISonarrJackettService;
 import com.lckp.jproxy.service.ISonarrProwlarrService;
-import com.lckp.jproxy.service.ISonarrQbittorrentService;
-import com.lckp.jproxy.service.ISonarrTransmissionService;
 
 import lombok.RequiredArgsConstructor;
 
@@ -39,10 +35,6 @@ public class FilterConfig {
 	private final IRadarrJackettService radarrJackettService;
 
 	private final IRadarrProwlarrService radarrProwlarrService;
-	
-	private final ISonarrQbittorrentService sonarrQbittorrentService;
-	
-	private final ISonarrTransmissionService sonarrTransmissionService;
 
 	@Bean
 	FilterRegistrationBean<SonarrJackettFilter> sonarrJackettFilter() {
@@ -77,24 +69,6 @@ public class FilterConfig {
 		bean.setFilter(new RadarrProwlarrFilter(radarrProwlarrService));
 		bean.setOrder(Integer.MIN_VALUE);
 		bean.addUrlPatterns(Common.FILTER_RADARR_PROWLARR_PATH);
-		return bean;
-	}
-
-	@Bean
-	FilterRegistrationBean<SonarrQbittorrentFilter> sonarrQbittorrentFilter() {
-		FilterRegistrationBean<SonarrQbittorrentFilter> bean = new FilterRegistrationBean<>();
-		bean.setFilter(new SonarrQbittorrentFilter(sonarrQbittorrentService));
-		bean.setOrder(Integer.MIN_VALUE);
-		bean.addUrlPatterns(Common.FILTER_SONARR_QBITTORRENT_PATH);
-		return bean;
-	}
-
-	@Bean
-	FilterRegistrationBean<SonarrTransmissionFilter> sonarrTransmissionFilter() {
-		FilterRegistrationBean<SonarrTransmissionFilter> bean = new FilterRegistrationBean<>();
-		bean.setFilter(new SonarrTransmissionFilter(sonarrTransmissionService));
-		bean.setOrder(Integer.MIN_VALUE);
-		bean.addUrlPatterns(Common.FILTER_SONARR_TRANSMISSION_PATH);
 		return bean;
 	}
 }