+++ src/main/java/com/lckp/jproxy/constant/Common.java
@@ -46,4 +46,6 @@ public class Common {
 	public static final String RULE_SYNC_AUTHORS_ALL = "ALL";
 	// 标题主规则 ID
 	public static final String MOST_IMPORTANT_TITLE_RULE_ID = "00000000000000000000000000000000";
+	// 视频文件扩展名正则表达式
+	public static final String VIDEO_EXTENSION_REGEX = "\\.(mp4|avi|wmv|flv|mov|mkv|webm|mpg|mpeg|3gp|iso)$";
 }