+++ src/main/java/com/lckp/jproxy/service/IQbittorrentService.java
@@ -1,0 +1,13 @@
+package com.lckp.jproxy.service;
+
+/**
+ * <p>
+ * qBittorrent 服务类
+ * </p>
+ *
+ * @author LuckyPuppy514
+ * @date 2023-04-11
+ */
+public interface IQbittorrentService extends IDownloaderService {
+
+}